+++ scripts/dataset_processing/tts/libriheavy/mfa_config.yaml
@@ -1,0 +1,3 @@
+punctuation: "、。।，@<>\"(),.:;¿?¡!\\&%#*~【】，…‥「」『』〝〟″⟨⟩♪・‹›«»～′$+=_°^[]{}"
+word_break_markers: "？！!()，,.:;¡¿?“„\"”&~%#—…‥、。【】$+=〝〟″‹›«»・⟨⟩「」『』-^"
+use_speaker_adaptation: true
